Calm before the (threaded) storm
The GUI now is properly organized, but it isn't functional yet. I'm thinking I need to implement some kind of threading, which won't be easy, as I've never worked with it before. I'm going to do my best with this. This is my checkpoint before I potentially mess up my code.